Avoid compiler warnings.

+++ apps/webserver/httpd.c
@@ -30,7 +30,7 @@
  *
  * Author: Adam Dunkels <adam@sics.se>
  *
- * $Id: httpd.c,v 1.2 2006/08/09 16:13:39 bg- Exp $
+ * $Id: httpd.c,v 1.3 2006/08/16 22:09:51 oliverschmidt Exp $
  */
 
 #include "contiki-net.h"
@@ -44,7 +44,7 @@
 #define STATE_WAITING 0
 #define STATE_OUTPUT  1
 
-#define SEND_STRING(s, str) PSOCK_SEND(s, str, strlen(str))
+#define SEND_STRING(s, str) PSOCK_SEND(s, str, (unsigned int)strlen(str))
 MEMB(conns, struct httpd_state, 8);
 
 #define ISO_nl      0x0a
